chore: Move test_db_migration to its own module and add it into git workflow
